publishing the website
publishing all of the files for the completed mvc website. The website just shows the contents of a database with a single table, courses, which holds info about courses such as name, number, and description. From the main page of the site, a user can add or remove courses to the database